Small change in dispose() of ZigBeeCoordinatorHandler. The condition for a deletion of the file are now checked in a different scope.

+++ org.openhab.binding.zigbee/src/main/java/org/openhab/binding/zigbee/handler/ZigBeeCoordinatorHandler.java
@@ -273,11 +273,12 @@ public abstract class ZigBeeCoordinatorHandler extends BaseBridgeHandler
 
             // Shut down the ZigBee library
             networkManager.shutdown();
-
-            if (bridgeRemoved) {
-                networkStateSerializer.delete();
-            }
         }
+
+        if (networkStateSerializer != null && bridgeRemoved) {
+            networkStateSerializer.delete();
+        }
+
         logger.debug("ZigBee network [{}] closed.", thing.getUID());
     }
